+++ src/Cwaff-Misc/CwaffSynergies.cs
@@ -190,6 +190,8 @@ public static class CwaffSynergies
             .MultFireRate(2.0f);
         // Ki Blast can be charged to fire a Kamehameha capable of breaking boss damage caps. Charging a Kamehameha slows the player down, and firing the Kamehameha prevents the player from moving entirely.
         NewMastery<MasteryOfKiBlast>(MASTERY_KI_BLAST, KiBlast.ItemName);
+        // Hallaeribut becomes permanently ravenous and can be fed items on the ground by reloading with a full clip. Items grant ammo proportional to their quality. If Hallaeribut runs out of ammo, it will automatically consume the least valuable item in the player's inventory for ammo. Hallaeribut will not feed on the player until no more items are available.
+        NewMastery<MasteryOfHallaeribut>(MASTERY_HALLAERIBUT, Hallaeribut.ItemName);
       #endregion
 
         SanityCheckAllSynergiesHaveBeenInitialized();
@@ -426,6 +428,7 @@ internal class MasteryOfAlienNailgun    : MasteryDummyItem {}
 internal class MasteryOfVladimir        : MasteryDummyItem {}
 internal class MasteryOfMaestro         : MasteryDummyItem {}
 internal class MasteryOfKiBlast         : MasteryDummyItem {}
+internal class MasteryOfHallaeribut     : MasteryDummyItem {}
 
 public enum Synergy {
     // Synergies
@@ -504,5 +507,6 @@ public enum Synergy {
     MASTERY_ALIEN_NAILGUN,
     MASTERY_VLADIMIR,
     MASTERY_MAESTRO,
-    MASTERY_KI_BLAST
+    MASTERY_KI_BLAST,
+    MASTERY_HALLAERIBUT,
 };